pyproject.toml
[project]
name = "ocr-space-mcp"
version = "0.1.0"
description = "MCP server for OCR.space API - text extraction from images and PDFs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Björn Büdenbender", email = "agentic.ai.forge@gmail.com" }
]
keywords = ["mcp", "ocr", "ocr.space", "text-extraction", "pdf", "image"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pypdf>=5.0.0",
    "Pillow>=11.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
]

[project.scripts]
ocr-space-mcp = "ocr_space_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_space_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "PLR", # pylint refactoring
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "PLR2004", # magic value comparison (too noisy)
]

[tool.ruff.lint.pylint]
max-args = 10
max-statements = 50

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["src/ocr_space_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
